[rename/#153] Member 모델 패키지 이름 수정
