+++ app/src/main/java/com/bongtu/baekseo/data/mapper/MemberMapper.kt
@@ -1,7 +1,7 @@
 package com.bongtu.baekseo.data.mapper
 
 import com.bongtu.baekseo.data.dto.member.ProfileInfoDto
-import com.bongtu.baekseo.data.model.profile.ProfileInfo
+import com.bongtu.baekseo.data.model.member.ProfileInfo
 
 fun ProfileInfoDto.toModel() = ProfileInfo(
     memberName = memberName,
